Move high-level API to managed.ts
