+++ src/transpilation/utils.ts
@@ -1,10 +1,12 @@
 import { FileSystem } from "enhanced-resolve";
+import * as fs from "fs";
 import * as path from "path";
 import * as resolve from "resolve";
 import { SourceNode } from "source-map";
 import * as ts from "typescript";
 // TODO: Don't depend on CLI?
 import * as cliDiagnostics from "../cli/diagnostics";
+import { CompilerOptions } from "../CompilerOptions";
 import * as lua from "../LuaAST";
 import * as diagnosticFactories from "./diagnostics";
 
@@ -73,4 +75,46 @@ export function resolvePlugin(
     }
 
     return { result };
+}
+
+const libCache = new Map<string, ts.SourceFile>();
+export function createVirtualProgram(input: Record<string, string>, options: CompilerOptions = {}): ts.Program {
+    function notImplemented(): never {
+        throw new Error("Not implemented");
+    }
+
+    const getFileFromInput = (fileName: string) =>
+        input[fileName] ?? (fileName.startsWith("/") ? input[fileName.slice(1)] : undefined);
+
+    const compilerHost: ts.CompilerHost = {
+        useCaseSensitiveFileNames: () => false,
+        getCanonicalFileName: fileName => fileName,
+        getCurrentDirectory: () => "/",
+        fileExists: fileName => fileName.startsWith("lib.") || getFileFromInput(fileName) !== undefined,
+        readFile: notImplemented,
+        writeFile: notImplemented,
+        getDefaultLibFileName: ts.getDefaultLibFileName,
+        getNewLine: () => "\n",
+
+        getSourceFile(fileName) {
+            const fileFromInput = getFileFromInput(fileName);
+            if (fileFromInput !== undefined) {
+                return ts.createSourceFile(fileName, fileFromInput, ts.ScriptTarget.Latest, false);
+            }
+
+            if (libCache.has(fileName)) return libCache.get(fileName)!;
+
+            if (fileName.startsWith("lib.")) {
+                const typeScriptDir = path.dirname(require.resolve("typescript"));
+                const filePath = path.join(typeScriptDir, fileName);
+                const content = fs.readFileSync(filePath, "utf8");
+
+                const sourceFile = ts.createSourceFile(fileName, content, ts.ScriptTarget.Latest, false);
+                libCache.set(fileName, sourceFile);
+                return sourceFile;
+            }
+        },
+    };
+
+    return ts.createProgram(Object.keys(input), options, compilerHost);
 }
